requirements.txt
python-docx>=0.8.11
beautifulsoup4>=4.9.3
lxml>=4.6.3 